requirements.txt
Django
django-markup
django-taggit
docutils
Pygments
